Migrated some tests to separate file
